New file checkSystemResources.py
checkSystemResources.py contains functions to check the current system's
various resources. check_disk_usage checks the total and free disk
space, which gives a warning if below 20 percent of the total space
free, and displays the total and free disk space in GB. check_cpu_usage
checks the usage of the cpu over an interval of 1 second and gives a
warning if the current cpu usage is over 75 percent. It also displays
the current number of physical cpu cores as well as the current logical
cpu cores. check_memory_usage checks and displays the system's current
virtual memory (excluding swap) in GB as well as displays a warning if
the available memory (excluding swap) is lower than 1 GB.
check_reboot_required looks for the file '/var/run/reboot-required' to
determine if a system reboot is required and will output a message if a
reboot is required or not.