Add 3 new functions to test the systems network
The function check_localhost uses the gethostbyname function provided by
the socket module to check if the 'localhost' host name translates to
the proper IPv4 address of '127.0.0.1'. The function check_internet
takes 3 parameters (IPv4 address, port number, and an integer) as a host,
port, and timeout value, respectively. check_internet makes a connection
attempt for the IPv4 address of '8.8.8.8' (public Google DNS server) on
port 53 (DNS) to test if the system can reach outside the local network.
The function check_DNS takes a single parameter (str) as a host name
which then uses the gethostbyname function provided by the socket module
to check if 'google-public-dns-a.google.com' (public Google DNS server)
correctly translates to the IPv4 address of '8.8.8.8'.