Add main function and __name__ == "__main__"
Added the main() function to control what code is run in the program
as well as the `if __name__ == "__main__"` conditional to better control
how the program is handled when executed as a script and when the
program is imported. Also moved any variables that are constant to after
the imports and before the function definitions.