Add new file regexSearch.py
The program regexSearch.py uses search_files to walk through a given
directory path, open any .txt files, and then search the opened file for
a given regex pattern. If the pattern is found, the file is appended to
the matchedFiles list. search_files only searches the given directory
and ignores any subdirectories. The function output_list takes the
matchedFiles list and prints out each file that contained the given
regex.

The program is checked to ensure that 2 command-line arguments were
given when the program was called. It also checks to ensure that the
given directory exists before taking any actions.