Add new script findImages.py
Takes a single command-line argument ( the path to a directory)
and uses the Image class in the Pillow module to open each file
in the specified directory to determine if it is an image. If the
file is not an image, the loop continues via an IOError exception
and skips the file.