Add new function and modify existing functions
Added new function generate_report to generate a PDF file report using
the reportlab module. Creates the title for the report from the date and
time (HH:MM:SS) from when the script is executed. Changed the existing
functions to append their results to a list and return that list to a
main list of data that will be passed to the generate_report function
after all the test/checks are passed.