Add a file named playChess.py
The program playChess.py allows two players to play a game of text-based
chess on the command-line. The program can create a new game from
scratch and save it as a json file with the player's names. Saved games
can also be loaded from a json file to continue a previous game.