Add bug fix and file formatting
Fixed a bug in the loadBoard function that attempted to look for a file
in a directory that was unknown if it existed. Added an os.path.isdir()
conditional to check for the directory and create it if it doesn't
already exist.

Added the datetime module to provide a formatted timestamp of the
date when the chess game is saved. The format for the timestamp is in
mm-dd-yy and is appended to the end of the chess file.