Add functionality to loadBoard function
Made changes to the loadBoard function to create a second dictionary
that tracks the number of moves each player has made. The number of
moves made by each player will be saved into the chess file so that the
player who made the latest move before the game was previously saved if
a file is loaded.