Add function chessTurns
Added a new function chessTurns which takes 4 arguments of
chessBoard (dic), playerMoves (dic), playerWhite (str) and
playerBlack (str). This function is in an infinite loop until the
saveBoard function is called. The function allows the two players to
take turns moving pieces around on the chess board and ensures that the
players do not interact with a space that is not a valid chess board
space.

The player turn is determined based on the number of moves each player
has taken. If the moves are equal, then White takes a turn, but if White
has more moves than Black, then Black takes a turn. The program exits
with an error if Black has more moves than White since that is not
possible since White always moves first in a chess game.

The players can either move a piece each turn or save and exit the
program on each turn. Moving a piece allows the player to type in the
name of a space on the board to select the piece that they would like
to move, then the player selects the space that they would like to move
that piece to. Selecting the save and quit option calls the saveBoard
function which saves the current board configuration and number of moves
for each player to a .json file.