Change chess board space names
This edit focuses on changing the chess board space names used in the
program such as changing '7a' to 'a7'. The edit also changes the layout
of the printed chess board so that the white pieces are at the bottom
and the row numbering is changed so that it starts at 1 from the white
side of the board.