Add a function pawn_promotion
The function pawn_promotion takes 3 arguments: the chessBoard, the
destination space that the player has chosen, and the colored pieces
of the player. A check is made in the chessTurns function that checks
if a white or black pawn has made it to its final rank (the other end
of the chess board from the side that it started on) and, if so, calls
the pawn_promotion function.

The pawn_promotion function asks the player what they would like to
promote their pawn to, adds the selected piece to the board in the same
space that the pawn was in, and then returns the chessBoard variable
back to the chessTurns function. After the pawn_promotion finishes, a
pawn piece is added to the removed pieces list for whatever color piece
the player was using.