Edit load_board function
Edited the load_board function to check if the chess directory is empty
before asking the players if they have a saved game. This allows the
program to load the default chess board if there are no saved games
without requiring user interaction. Made saved game selection easier
by implementing a menu of the different file options with pyinputplus.