Add backup-files.sh bash script
The backup-files.sh bash script is used to run the backupFile,py script.
The backup-files.sh script takes no arguments and checks if the backupFile.py
script (1) exists in the path specified in the shell script, (2) is a regular file
and not a directory or a device file, and (3) if the backupFile.py script has execute
permissions for the current user.